added MessageServiceImpl.java and Favourite.java, changed Appointment.java and MessageService.java
